+++ Src/Game/src/app/app.module.ts
@@ -1,26 +1,27 @@
 import { BrowserModule } from "@angular/platform-browser";
 import { NgModule } from "@angular/core";
 import { AppComponent } from "./app.component";
-import { RouterModule } from "@angular/router";
-import { routes } from "./app.routes";
+import { AppRoutingModule } from "./app.routing.module";
 import { HttpClient, HttpClientModule } from "@angular/common/http";
 import { HttpLoaderFactory } from "./app.module.factories";
 import { TranslateLoader, TranslateModule } from "@ngx-translate/core";
 import { ServicesModule } from "./shared/services/services.module";
 import { LoaderModule } from "./loader/loader.module";
 import { GameModule } from "./game/game.module";
+import { MenuModule } from "./menu/menu.module";
 
 @NgModule({
   declarations: [
     AppComponent
   ],
   imports: [
-    BrowserModule,
     GameModule,
-    HttpClientModule,
+    MenuModule,
     LoaderModule,
-    RouterModule.forRoot(routes),
+    BrowserModule,
     ServicesModule,
+    AppRoutingModule,
+    HttpClientModule,
     TranslateModule.forRoot({
       loader: {
         provide: TranslateLoader,
