+++ Src/Game/src/app/loader/loader.component.html
@@ -9,6 +9,7 @@
     <p [translate]="'states.loading'"></p>
 
     <!-- Progress bar. -->
+    <a (click)="go()">Menu</a>
 
   </div>
