+++ Src/Game/src/app/menu/menu.component.html
@@ -6,6 +6,9 @@
     <h2 [translate]="'titles.gameName'"></h2>
 
     <!-- Menu options. -->
+    <nav>
+      <button class="red">Stats</button>
+    </nav>
 
   </div>
