+++ Src/Game/src/colors.scss
@@ -1,1 +1,4 @@
+/* Pocket Investigator */
+
 $red: #ff0000;
+$white: #ffffff;
