+++ Src/Game/src/styles.scss
@@ -21,4 +21,8 @@ body {
   height: 100%;
   display: flex;
   flex-direction: column;
+
+  > ng-component {
+    height: 100%;
+  }
 }
